requirements.txt
asttokens==3.0.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
executing==2.1.0
filelock==3.13.1
frozendict==2.4.6
fsspec==2024.2.0
h11==0.14.0
html5lib==1.1
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==4.9.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multitasking==0.0.11
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.8
pillow==10.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pyzmq==26.2.0
requests==2.32.3
requests-futures==1.0.2
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.27.1
setuptools==70.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
yahooquery==2.3.7
yfinance==0.2.50
